PR resolve
- Eslint & Prettier
- fixing imports
- Separated context into AppContext and LoginContext
- global Colors.scss file
