+++ nextjs-nabil/pages/homepage.js
@@ -1,10 +1,11 @@
 import React, { useContext, useEffect } from "react";
 import { useRouter } from "next/router";
-import { AppContext } from "../src/context";
+// import { AppContext } from "../src/context";
 import ViewGames from "../src/views/ViewGames";
+import { LoginContext } from "../src/LoginContext";
 
 export default function homepage() {
-  const {  isAuthenticated } = useContext(AppContext);
+  const { isAuthenticated } = useContext(LoginContext);
   const router = useRouter();
 
   useEffect(() => {
