+++ nextjs-nabil/pages/index.js
@@ -1,11 +1,12 @@
-import { AppContext } from "../src/context";
+// import { AppContext } from "../src/context";
 import { useRouter } from "next/router";
-import { useContext, useEffect } from "react";
+import React, { useContext, useEffect } from "react";
 import { Box } from "@mui/system";
 import Logincard from "../src/components/logincard";
+import { LoginContext } from "../src/LoginContext";
 
 export default function Home() {
-  const { isAuthenticated } = useContext(AppContext);
+  const { isAuthenticated } = useContext(LoginContext);
   const router = useRouter();
 
   useEffect(() => {
