+++ nextjs-nabil/pages/teams.js
@@ -1,10 +1,10 @@
 import { useContext, useEffect } from "react";
 import { useRouter } from "next/router";
-import { AppContext } from "../src/context";
 import ViewTeams from "../src/views/ViewTeams";
+import { LoginContext } from "../src/LoginContext";
 
 export default function homepage() {
-  const { isAuthenticated } = useContext(AppContext);
+  const { isAuthenticated } = useContext(LoginContext);
   const router = useRouter();
 
   useEffect(() => {
