+++ nextjs-nabil/src/components/CompBottomNavigation.js
@@ -1,5 +1,4 @@
-// import * as React from "react";
-import { useState, useContext } from "react";
+import React, { useContext } from "react";
 import Box from "@mui/material/Box";
 import BottomNavigation from "@mui/material/BottomNavigation";
 import BottomNavigationAction from "@mui/material/BottomNavigationAction";
@@ -7,7 +6,7 @@ import GroupsIcon from "@mui/icons-material/Groups";
 import PersonIcon from "@mui/icons-material/Person";
 import SportsBasketballIcon from "@mui/icons-material/SportsBasketball";
 import { useRouter } from "next/router";
-import { AppContext } from "../context";
+import { AppContext } from "../AppContext";
 
 const CompBottomNavigation = () => {
   const { valueBottomNavbar, setvalueBottomNavbar } = useContext(AppContext);
@@ -16,7 +15,7 @@ const CompBottomNavigation = () => {
     <div>
       <Box
         sx={{
-          width: 500,
+          // width: 500,
           display: { xs: "initial", md: "none" },
           position: "",
           textAlign: "center",
