+++ nextjs-nabil/src/components/PopupModal.js
@@ -1,24 +1,11 @@
-import * as React from "react";
-import { useContext, useState } from "react";
-import { AppContext } from "../context";
+import React, { useContext, useState } from "react";
+import { AppContext } from "../AppContext";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import Modal from "@mui/material/Modal";
 import LoadingComp from "./LoadingComp";
 import styles from "../../styles/modal.module.scss";
-
-const style = {
-  position: "absolute",
-  top: "50%",
-  left: "50%",
-  transform: "translate(-50%, -50%)",
-  width: 400,
-  bgcolor: "background.paper",
-  border: "2px solid #000",
-  boxShadow: 24,
-  p: 4,
-};
 
 const PopupModal = ({ index }) => {
   const [open, setOpen] = useState(false);
