+++ nextjs-nabil/src/layouts/Layout.js
@@ -1,12 +1,11 @@
 import React from "react";
 import { useContext } from "react";
-import { AppContext } from "../context";
-
 import Navbar from "../components/NavBar";
 import Footer from "../components/Footer";
+import { LoginContext } from "../LoginContext";
 
 const Layout = ({ children }) => {
-  const { isAuthenticated } = useContext(AppContext);
+  const { isAuthenticated } = useContext(LoginContext);
 
   if (isAuthenticated === "true") {
     return (
