+++ nextjs-nabil/styles/Colors.scss
@@ -1,0 +1,8 @@
+
+// Original Values
+// $primarycolor: #a94949;
+// $secondarycolor: #3474ab;
+
+
+$primarycolor: #6568a5;
+$secondarycolor: hsl(237, 50%, 52%);
