+++ nextjs-nabil/styles/footer.module.scss
@@ -1,4 +1,4 @@
-$primarycolor: #a94949;
+@import 'Colors';
 
 .footer {
   padding: 0.5px;
@@ -8,5 +8,7 @@ $primarycolor: #a94949;
   .h1 {
     color: rgb(251, 251, 251);
     text-align: center;
+    font-weight: normal;
+    font-size: x-large;
   }
 }
