+++ nextjs-nabil/styles/globals.scss
@@ -1,7 +1,8 @@
 @import 'https://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700';
+@import 'Colors';
 $open-sans: 'Open Sans', sans-serif;
 $fontfamily: $open-sans;
-$primarycolor: #a94949;
+
 
 html,
 body {
