+++ nextjs-nabil/styles/loginpage.module.scss
@@ -1,4 +1,4 @@
-$primarycolor: #a94949;
+@import 'Colors';
 
 .logincard {
   .button {
