+++ nextjs-nabil/styles/modal.module.scss
@@ -1,4 +1,4 @@
-$primarycolor: #a94949;
+@import 'Colors';
 
 .button {
   background-color: $primarycolor;
