+++ nextjs-nabil/styles/navbar.module.scss
@@ -1,5 +1,4 @@
-$primarycolor: #a94949;
-$secondarycolor: #3474ab;
+@import 'Colors';
 
 .toolbar {
   background-color: $primarycolor;
